+++ lib/govuk_publishing_components/presenters/contextual_navigation.rb
@@ -26,15 +26,22 @@ module GovukPublishingComponents
       end
 
       def taxon_breadcrumbs
-        @taxon_breadcrumbs ||= ContentBreadcrumbsBasedOnTaxons.new(content_item).breadcrumbs
+        @taxon_breadcrumbs ||= ContentBreadcrumbsBasedOnTaxons.call(content_item)
       end
 
       def priority_breadcrumbs
         @priority_breadcrumbs ||= ContentBreadcrumbsBasedOnPriority.call(content_item)
       end
 
+      def topic_breadcrumbs
+        @topic_breadcrumbs ||= ContentBreadcrumbsBasedOnTopic.call(content_item)
+      end
+
       def breadcrumbs
-        return breadcrumbs_based_on_parent unless content_tagged_to_a_finder?
+        breadcrumbs_based_on_ancestors
+      end
+
+      def finder_breadcrumbs
         return [] unless parent_finder
 
         [
@@ -75,6 +82,10 @@ module GovukPublishingComponents
 
       def content_is_a_html_publication?
         content_item["document_type"] == "html_publication"
+      end
+
+      def content_has_a_topic?
+        content_item.dig("links", "topics").present?
       end
 
       def tagged_to_brexit?
@@ -119,8 +130,8 @@ module GovukPublishingComponents
         step_nav_helper.show_also_part_of_step_nav?
       end
 
-      def breadcrumbs_based_on_parent
-        ContentBreadcrumbsBasedOnParent.new(content_item).breadcrumbs[:breadcrumbs]
+      def breadcrumbs_based_on_ancestors
+        ContentBreadcrumbsBasedOnAncestors.call(content_item)
       end
 
       def step_nav_helper